+++ demo-app/changelog.md
@@ -11,11 +11,12 @@ The `svg-dom-demo` gallery is not published as part of the crate's release, so t
 ### Added
 
 - Make `feTurbulence`/`feDisplacementMap` demo interactive (`0088502`)
-- Add specific turbulence scale 0 render test (``)
+- Add specific turbulence scale 0 render test (`17d0f1d`)
 
 ### Fixed
 
 - Doc only: correct description of `scale` (`6005461`)
+- Label channel slider positions (``)
 
 ## [0.1.12] - 2026-08-04